Add geolocation features and update UI components for location display , update the UI for the home page , the location does not work
