+++ .vscode/settings.json
@@ -2,6 +2,7 @@
   "cSpell.words": ["azrieli"],
   // "cmake.sourceDirectory": "C:/Users/GIGABYTE/Desktop/final_project/smart_guide/linux",
   "java.configuration.updateBuildConfiguration": "interactive",
+  "cmake.ignoreCMakeListsMissing": true,
   //"cmake.ignoreCMakeListsMissing": true
   
 }
