+++ smart_guide/lib/Screens/login_screen.dart
@@ -47,11 +47,10 @@ class _LoginScreenState extends State<LoginScreen> {
                     crossAxisAlignment: CrossAxisAlignment.center,
                     children: [
                       //Heading Text
-                      HeadingText('Welcome to Smart Guide'),
-
+                      HeadingText('Welcome Back!', 40),
                       // Body Text
-                      // BodyText(
-                      //     text: 'welcome back we missed you', fontSize: 15),
+                      BodyText(
+                          text: 'welcome back we missed you', fontSize: 15),
                       const SizedBox(height: 40),
                       // username text field
                       // Aligns the text to the left
