+++ smart_guide/pubspec.yaml
@@ -20,6 +20,8 @@ dependencies:
   english_words: ^4.0.0
   provider: ^6.0.0
   firebase_auth: ^5.3.4
+  geolocator: ^10.1.0
+  geocoding: ^2.1.0
   # url_launcher: ^6.0.20
   # english_words: ^4.0.0
   # provider: ^6.0.0
